remove get_formset_data method, move create_condition_and_components to MeasureCreationPattern
